refactored VBSS tests
