+++ src/test/java/org/cicirello/search/ss/VBSSIntCostTests.java
@@ -27,8 +27,8 @@ import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.SolutionCostPair;
 import org.junit.jupiter.api.*;
 
-/** JUnit tests for ValueBiasedStochasticSampling. */
-public class VBSSTests extends SharedTestStochasticSampler {
+/** JUnit tests for ValueBiasedStochasticSampling problems with int-valued costs. */
+public class VBSSIntCostTests extends SharedTestStochasticSampler {
 
   @Test
   public void testExponentialBiasFunction() {
@@ -37,6 +37,57 @@ public class VBSSTests extends SharedTestStochasticSampler {
     double[] expected = {1, Math.exp(1), Math.exp(2), Math.exp(3)};
     for (int v = 0, i = 0; v <= 12; v += 4, i++) {
       assertEquals(expected[i], bias.bias(v), 1E-10);
+    }
+  }
+
+  @Test
+  public void testAdjustForBias() {
+    for (int n = 2; n < 8; n++) {
+      IntProblem problem = new IntProblem();
+      IntHeuristic h = new IntHeuristic(problem, n);
+      ValueBiasedStochasticSampling<Permutation> ch =
+          new ValueBiasedStochasticSampling<Permutation>(h);
+      for (int k = 2; k <= n; k++) {
+        double[] values = new double[n];
+        for (int i = 0, j = 1; i < k; i++, j *= 2) {
+          values[k - 1 - i] = j;
+        }
+        ch.adjustForBias(values, k);
+        assertEquals(1.0, values[k - 1], 1E-10);
+        for (int i = k - 3; i >= 0; i--) {
+          assertEquals(2 * (values[i + 2] - values[i + 1]), values[i + 1] - values[i], 1E-10);
+        }
+      }
+    }
+  }
+
+  @Test
+  public void testSelect() {
+    for (int n = 2; n < 8; n++) {
+      IntProblem problem = new IntProblem();
+      IntHeuristic h = new IntHeuristic(problem, n);
+      ValueBiasedStochasticSampling<Permutation> ch =
+          new ValueBiasedStochasticSampling<Permutation>(h);
+      for (int k = 2; k <= n; k++) {
+        double inc = 1.0 / k;
+        double[] values = new double[n];
+        values[0] = inc;
+        for (int i = 1; i < k; i++) {
+          values[i] = values[i - 1] + inc;
+        }
+        double u = 0.0;
+        for (int i = 0; i < k; i++, u += inc) {
+          assertEquals(i, ch.select(values, k, u));
+        }
+        u = inc / 2;
+        for (int i = 0; i < k; i++, u += inc) {
+          assertEquals(i, ch.select(values, k, u));
+        }
+        u = 1.0 - 1E-10;
+        for (int i = k - 1; i >= 0; i--, u -= inc) {
+          assertEquals(i, ch.select(values, k, u));
+        }
+      }
     }
   }
 
@@ -93,32 +144,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCosts() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize();
-      assertEquals(2, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
   public void testWithIntCostsMultipleSamples() {
     for (int n = 0; n < 10; n++) {
       IntProblem problem = new IntProblem();
@@ -145,32 +170,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCostsMultipleSamples() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize(5);
-      assertEquals(5, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize(2);
-      assertEquals(7, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
   public void testWithIntCostsWithProgressTracker() {
     for (int n = 0; n < 10; n++) {
       ProgressTracker<Permutation> originalTracker = new ProgressTracker<Permutation>();
@@ -186,30 +185,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
       assertEquals(1, ch.getTotalRunLength());
       assertEquals((n + 1) * n / 2, solution.getCost());
       assertEquals((n + 1) * n / 2, tracker.getCost());
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testWithDoubleCostsWithProgressTracker() {
-    for (int n = 0; n < 10; n++) {
-      ProgressTracker<Permutation> originalTracker = new ProgressTracker<Permutation>();
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h, originalTracker);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      assertTrue(originalTracker == tracker);
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
       Permutation p = solution.getSolution();
       assertEquals(n, p.length());
       tracker = new ProgressTracker<Permutation>();
@@ -242,80 +217,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCostsSplit() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> chOriginal =
-          new ValueBiasedStochasticSampling<Permutation>(h);
-      ValueBiasedStochasticSampling<Permutation> ch = chOriginal.split();
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testAdjustForBias() {
-    for (int n = 2; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h);
-      for (int k = 2; k <= n; k++) {
-        double[] values = new double[n];
-        for (int i = 0, j = 1; i < k; i++, j *= 2) {
-          values[k - 1 - i] = j;
-        }
-        ch.adjustForBias(values, k);
-        assertEquals(1.0, values[k - 1], 1E-10);
-        for (int i = k - 3; i >= 0; i--) {
-          assertEquals(2 * (values[i + 2] - values[i + 1]), values[i + 1] - values[i], 1E-10);
-        }
-      }
-    }
-  }
-
-  @Test
-  public void testSelect() {
-    for (int n = 2; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h);
-      for (int k = 2; k <= n; k++) {
-        double inc = 1.0 / k;
-        double[] values = new double[n];
-        values[0] = inc;
-        for (int i = 1; i < k; i++) {
-          values[i] = values[i - 1] + inc;
-        }
-        double u = 0.0;
-        for (int i = 0; i < k; i++, u += inc) {
-          assertEquals(i, ch.select(values, k, u));
-        }
-        u = inc / 2;
-        for (int i = 0; i < k; i++, u += inc) {
-          assertEquals(i, ch.select(values, k, u));
-        }
-        u = 1.0 - 1E-10;
-        for (int i = k - 1; i >= 0; i--, u -= inc) {
-          assertEquals(i, ch.select(values, k, u));
-        }
-      }
-    }
-  }
-
-  @Test
   public void testWithIntCostsExponent() {
     for (int n = 0; n < 10; n++) {
       IntProblem problem = new IntProblem();
@@ -335,32 +236,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
       assertEquals(2, ch.getTotalRunLength());
       assertEquals((n + 1) * n / 2, solution.getCost());
       assertEquals((n + 1) * n / 2, tracker.getCost());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testWithDoubleCostsExponent() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h, 2.0);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize();
-      assertEquals(2, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
       tracker = new ProgressTracker<Permutation>();
       ch.setProgressTracker(tracker);
       assertTrue(tracker == ch.getProgressTracker());
@@ -392,30 +267,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCostsExponentWithProgressTracker() {
-    for (int n = 0; n < 10; n++) {
-      ProgressTracker<Permutation> originalTracker = new ProgressTracker<Permutation>();
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h, 2.0, originalTracker);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      assertTrue(originalTracker == tracker);
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
   public void testWithIntCostsExponentSplit() {
     for (int n = 0; n < 10; n++) {
       IntProblem problem = new IntProblem();
@@ -430,29 +281,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
       assertEquals(1, ch.getTotalRunLength());
       assertEquals((n + 1) * n / 2, solution.getCost());
       assertEquals((n + 1) * n / 2, tracker.getCost());
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testWithDoubleCostsExponentSplit() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> chOriginal =
-          new ValueBiasedStochasticSampling<Permutation>(h, 2.0);
-      ValueBiasedStochasticSampling<Permutation> ch = chOriginal.split();
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
       Permutation p = solution.getSolution();
       assertEquals(n, p.length());
       tracker = new ProgressTracker<Permutation>();
@@ -495,39 +323,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCostsBias() {
-    ValueBiasedStochasticSampling.BiasFunction bias =
-        new ValueBiasedStochasticSampling.BiasFunction() {
-          @Override
-          public double bias(double value) {
-            return value * value;
-          }
-        };
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h, bias);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize();
-      assertEquals(2, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
   public void testWithIntCostsBiasWithProgressTracker() {
     ValueBiasedStochasticSampling.BiasFunction bias =
         new ValueBiasedStochasticSampling.BiasFunction() {
@@ -550,37 +345,6 @@ public class VBSSTests extends SharedTestStochasticSampler {
       assertEquals(1, ch.getTotalRunLength());
       assertEquals((n + 1) * n / 2, solution.getCost());
       assertEquals((n + 1) * n / 2, tracker.getCost());
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testWithDoubleCostsBiasWithProgressTracker() {
-    ValueBiasedStochasticSampling.BiasFunction bias =
-        new ValueBiasedStochasticSampling.BiasFunction() {
-          @Override
-          public double bias(double value) {
-            return value * value;
-          }
-        };
-    for (int n = 0; n < 10; n++) {
-      ProgressTracker<Permutation> originalTracker = new ProgressTracker<Permutation>();
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n);
-      ValueBiasedStochasticSampling<Permutation> ch =
-          new ValueBiasedStochasticSampling<Permutation>(h, bias, originalTracker);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      assertTrue(originalTracker == tracker);
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
       Permutation p = solution.getSolution();
       assertEquals(n, p.length());
       tracker = new ProgressTracker<Permutation>();
